+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 1.29.1
+
+### Fixed
+
+- Exception on Gradle 8.13:
+  ```
+  java.lang.NoSuchMethodError: 'org.gradle.jvm.toolchain.internal.SpecificInstallationToolchainSpec org.gradle.jvm.toolchain.internal.SpecificInstallationToolchainSpec.fromJavaExecutable(org.gradle.api.model.ObjectFactory, java.lang.String)'
+  ```
+
 ## 1.29.0
 
 - `MpsCheck`, `MpsGenerate`, `MpsExecute`: put MPS jars at the beginning of the classpath of the corresponding backends,